+++ index.js
@@ -14,16 +14,16 @@ const authorityRoute = require("./routes/authority");
 const dashboardCollectionRoute = require("./routes/dashboard");
 const favUserRoute = require("./routes/favUser");
 const meetRoute = require("./routes/meet");
+const relationsRoute = require("./routes/relations");
 const otherRoute = require("./routes/other");
 const paymentRoute = require("./routes/payment");
 const planRoute = require("./routes/plan");
 const reviewRoute = require("./routes/review");
 const userRoute = require("./routes/user");
 const userVerificationRoute = require("./routes/userVerification");
-const { connectMongoClient, mongoClient } = require('./mongodbConnection');
-const mongoose = require('./mongooseConnection');
+const { connectMongoClient, mongoClient } = require("./mongodbConnection");
+const mongoose = require("./mongooseConnection");
 connectMongoClient();
-
 
 // middleware
 app.use(cors());
@@ -36,8 +36,9 @@ app.use("/", blogRoute);
 app.use("/", authorityRoute);
 
 app.use("/", dashboardCollectionRoute);
-app.use("/", favUserRoute);
-app.use("/", meetRoute);
+app.use("/", favUserRoute.router);
+app.use("/", meetRoute.router);
+app.use("/", relationsRoute);
 app.use("/", otherRoute);
 app.use("/", paymentRoute);
 app.use("/", planRoute);
@@ -45,9 +46,7 @@ app.use("/", reviewRoute);
 app.use("/", userRoute);
 app.use("/", userVerificationRoute);
 
-
 const uri = `mongodb+srv://${process.env.DB_User}:${process.env.DB_Pass}@cluster0.ymw1jdy.mongodb.net/?retryWrites=true&w=majority&appName=AtlasApp`;
-
 
 // DB_User = SoulMate-Matrimony
 // DB_Pass = LV2hgni1aq9w6d5H
@@ -55,7 +54,6 @@ const uri = `mongodb+srv://${process.env.DB_User}:${process.env.DB_Pass}@cluster
 // SSLID=soulm64e6111916384
 // SSLPASS=soulm64e6111916384@ssl
 // Create a MongoClient with a MongoClientOptions object to set the Stable API version
-
 
 // async function run() {
 //   try {
